# Redesigning visit_primitive so that it is optimized for strings and numbers (DOM backend)